questionController pagination

+++ src/main/java/com/javamentor/qa/platform/service/impl/dto/QuestionDtoServiceImpl.java
@@ -35,7 +35,7 @@ public class QuestionDtoServiceImpl implements QuestionDtoService {
         pageDto.setItems(questionDtoList);
         pageDto.setTotalResultCount(totalResultCount);
         pageDto.setCurrentPageNumber(page);
-        pageDto.setTotalPageCount((int) Math.ceil(totalResultCount / size));
+        pageDto.setTotalPageCount((int) Math.ceil(totalResultCount/(double)size));
         pageDto.setItemsOnPage(size);
 
         return pageDto;
